# api/openapi/releasepilot-v1.yaml
openapi: 3.1.0
info:
  title: ReleasePilot Control Plane API
  version: 0.1.0
  description: >-
    ReleasePilot MVP의 릴리스 요청, 승인, Canary 운영 명령,
    정책 및 감사 조회 API 계약이다.
servers:
  - url: /api/v1
    description: Same-origin API
...
paths:
...
  /releases:
    post:
      tags: [Releases]
      summary: 릴리스 요청 생성
      operationId: createRelease
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReleaseRequest'
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
    get:
      tags: [Releases]
      summary: 릴리스 목록 조회
      operationId: listReleases
      parameters:
        - name: serviceId
          in: query
          schema:
            type: string
            format: uuid
        - name: environmentId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReleaseStatus'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 릴리스 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleasePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /releases/{releaseId}:
    get:
      tags: [Releases]
      summary: 릴리스 상세 조회
      operationId: getRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
      responses:
        '200':
          description: 릴리스 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /releases/{releaseId}/promote:
    post:
      tags: [Operations]
      summary: 현재 Canary 단계를 다음 단계로 승격
      operationId: promoteRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/OperationReason'
      responses:
        '202':
          $ref: '#/components/responses/OperationAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /releases/{releaseId}/pause:
    post:
      tags: [Operations]
      summary: 실행 중인 릴리스 일시정지
      operationId: pauseRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/OperationReason'
      responses:
        '202':
          $ref: '#/components/responses/OperationAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /releases/{releaseId}/resume:
    post:
      tags: [Operations]
      summary: 일시정지된 릴리스 재개
      operationId: resumeRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/OperationReason'
      responses:
        '202':
          $ref: '#/components/responses/OperationAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /releases/{releaseId}/abort:
    post:
      tags: [Operations]
      summary: 릴리스 중단 및 롤백 요청
      operationId: abortRelease
      parameters:
        - $ref: '#/components/parameters/ReleaseId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/OperationReason'
      responses:
        '202':
          $ref: '#/components/responses/OperationAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: RELEASEPILOT_SESSION
  parameters:
...
    ReleaseId:
      name: releaseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 같은 사용자와 operation에서 24시간 동안 중복 실행을 방지한다.
      schema:
        type: string
        minLength: 16
        maxLength: 128
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  requestBodies:
    OperationReason:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OperationReasonRequest'
  responses:
    BadRequest:
      description: 요청 형식이 잘못됨
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 인증 필요
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: 권한 부족
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: 현재 상태 또는 동시 실행과 충돌
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableEntity:
      description: 형식은 유효하지만 의미 검증 실패
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    OperationAccepted:
      description: 비동기 운영 명령 접수됨
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OperationReceipt'
  schemas:
...
    CreateReleaseRequest:
      type: object
      additionalProperties: false
      required:
        - serviceId
        - environmentId
        - version
        - imageRepository
        - imageDigest
        - changeSummary
        - commitSha
        - pipelineUrl
      properties:
        serviceId:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        version:
          type: string
          minLength: 1
          maxLength: 100
        imageRepository:
          type: string
          minLength: 1
          maxLength: 500
        imageDigest:
          type: string
          pattern: '^sha256:[a-f0-9]{64}$'
        changeSummary:
          type: string
          minLength: 1
          maxLength: 2000
        commitSha:
          type: string
          pattern: '^[a-fA-F0-9]{40}$'
        pipelineUrl:
          type: string
          format: uri
        requestedPolicyVersionId:
          type: string
          format: uuid
    Release:
      type: object
      required:
        - id
        - serviceId
        - environmentId
        - version
        - imageDigest
        - status
        - requestedBy
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        version:
          type: string
        imageDigest:
          type: string
        status:
          $ref: '#/components/schemas/ReleaseStatus'
        requestedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        approvedAt:
          type: [string, 'null']
          format: date-time
        startedAt:
          type: [string, 'null']
          format: date-time
        finishedAt:
          type: [string, 'null']
          format: date-time
    ReleaseDetail:
      allOf:
        - $ref: '#/components/schemas/Release'
        - type: object
          required: [changeSummary, commitSha, pipelineUrl, steps]
          properties:
            changeSummary:
              type: string
            commitSha:
              type: string
            pipelineUrl:
              type: string
              format: uri
            policySnapshot:
              $ref: '#/components/schemas/PolicySnapshotSummary'
            approval:
              anyOf:
                - $ref: '#/components/schemas/ApprovalDecision'
                - type: 'null'
            steps:
              type: array
              items:
                $ref: '#/components/schemas/RolloutStep'
    ReleasePage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Release'
        nextCursor:
          type: [string, 'null']
    ReleaseStatus:
      type: string
      enum:
        - DRAFT
        - VALIDATING
        - VALIDATION_FAILED
        - PENDING_APPROVAL
        - APPROVED
        - REJECTED
        - CANCELED
        - DEPLOYING
        - ANALYZING
        - PROMOTING
        - PAUSED
        - ROLLING_BACK
        - ROLLED_BACK
        - ROLLBACK_FAILED
        - SUCCEEDED
        - FAILED
...
    OperationReasonRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 1000
...
    OperationReceipt:
      type: object
      required: [operationId, releaseId, status, acceptedAt]
      properties:
        operationId:
          type: string
          format: uuid
        releaseId:
          type: string
          format: uuid
        status:
          const: ACCEPTED
        acceptedAt:
          type: string
          format: date-time
    ApprovalDecision:
      type: object
      required: [decision, decidedBy, reason, decidedAt]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED]
        decidedBy:
          type: string
          format: uuid
        reason:
          type: string
        decidedAt:
          type: string
          format: date-time
    RolloutStep:
      type: object
      required: [index, weight, minimumObservationSeconds, status]
      properties:
        index:
          type: integer
          minimum: 0
        weight:
          type: integer
          minimum: 1
          maximum: 100
        minimumObservationSeconds:
          type: integer
        status:
          type: string
          enum: [PENDING, DEPLOYING, OBSERVING, EVALUATING, PASSED, FAILED, INCONCLUSIVE]
        decision:
          type: [string, 'null']
          enum: [PASS, FAIL, INCONCLUSIVE, null]
    PolicySnapshotSummary:
      type: object
      required: [sourcePolicyVersionId, name, checksum]
      properties:
        sourcePolicyVersionId:
          type: string
          format: uuid
        name:
          type: string
        checksum:
          type: string
...
    Problem:
      type: object
      description: RFC 9457 Problem Details
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        traceId:
          type: string
        violations:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
security:
  - sessionCookie: []