# apps/control-plane/src/main/java/kr/releasepilot/controlplane/release/api/ReleaseController.java
package kr.releasepilot.controlplane.release.api;
import jakarta.validation.Valid;import jakarta.validation.constraints.*;import kr.releasepilot.controlplane.identity.UserAccountPrincipal;import kr.releasepilot.controlplane.release.*;import org.hibernate.validator.constraints.URL;import org.springframework.http.*;import org.springframework.security.core.*;import org.springframework.security.core.annotation.AuthenticationPrincipal;import org.springframework.web.bind.annotation.*;import java.time.Instant;import java.util.UUID;
@RestController @RequestMapping("/api/v1/releases")public class ReleaseController{
 private final ReleaseRequestService service;private final ReleaseQueryService queries;public ReleaseController(ReleaseRequestService service,ReleaseQueryService queries){this.service=service;this.queries=queries;}
 @PostMapping @ResponseStatus(HttpStatus.CREATED)ReleaseResponse create(@RequestHeader("Idempotency-Key")@Size(min=16,max=128)String key,@Valid@RequestBody CreateReleaseRequest r,@AuthenticationPrincipal UserAccountPrincipal actor,Authentication authentication){var result=service.create(new ReleaseRequestService.Command(r.serviceId(),r.environmentId(),r.version(),r.imageRepository(),r.imageDigest(),r.changeSummary(),r.commitSha(),r.pipelineUrl(),r.requestedPolicyVersionId(),key),actor,authentication);return ReleaseResponse.from(result);}
 @GetMapping ReleasePage list(@RequestParam(defaultValue="20")int limit,Authentication auth){return new ReleasePage(queries.list(Math.clamp(limit,1,100),auth).stream().map(ReleaseResponse::from).toList(),null);}
 @GetMapping("/{id}") ReleaseDetail detail(@PathVariable UUID id,Authentication auth){return ReleaseDetail.from(queries.detail(id,auth));}
 public record CreateReleaseRequest(@NotNull UUID serviceId,@NotNull UUID environmentId,@NotBlank@Size(max=100)String version,@NotBlank@Size(max=500)String imageRepository,@Pattern(regexp="^sha256:[a-f0-9]{64}$")String imageDigest,@NotBlank@Size(max=2000)String changeSummary,@Pattern(regexp="^[a-fA-F0-9]{40}$")String commitSha,@NotBlank@URL@Size(max=1000)String pipelineUrl,UUID requestedPolicyVersionId){}
 public record ReleaseResponse(UUID id,UUID serviceId,UUID environmentId,String version,String imageDigest,String status,UUID requestedBy,Instant createdAt){static ReleaseResponse from(ReleaseRequestService.Result r){var v=r.release();return value(v,r.artifact().getImageDigest());}static ReleaseResponse from(ReleaseQueryService.Item item){return value(item.release(),item.imageDigest());}private static ReleaseResponse value(Release v,String digest){return new ReleaseResponse(v.getId(),v.getServiceId(),v.getEnvironmentId(),v.getVersion(),digest,v.getStatus().name(),v.getRequestedBy(),v.getCreatedAt());}}
 public record ReleasePage(java.util.List<ReleaseResponse> items,String nextCursor){}
 public record ReleaseDetail(UUID id,UUID serviceId,UUID environmentId,String version,String imageDigest,String status,UUID requestedBy,Instant createdAt,String changeSummary,String commitSha,String pipelineUrl,ReleaseQueryService.SnapshotSummary policySnapshot,ApprovalView approval,java.util.List<Object> steps){static ReleaseDetail from(ReleaseQueryService.Detail d){var value=d.item().release();return new ReleaseDetail(value.getId(),value.getServiceId(),value.getEnvironmentId(),value.getVersion(),d.item().imageDigest(),value.getStatus().name(),value.getRequestedBy(),value.getCreatedAt(),d.changeSummary(),d.commitSha(),d.pipelineUrl(),d.policySnapshot(),ApprovalView.from(d.approval()),java.util.List.of());}}
 public record ApprovalView(String decision,UUID decidedBy,String reason,Instant decidedAt){static ApprovalView from(kr.releasepilot.controlplane.approval.ApprovalDecision value){return value==null?null:new ApprovalView(value.getDecision().name(),value.getDecidedBy(),value.getReason(),value.getDecidedAt());}}
}