# apps/control-plane/src/main/java/kr/releasepilot/controlplane/rollout/RolloutOperationController.java
package kr.releasepilot.controlplane.rollout;
import jakarta.validation.Valid;import jakarta.validation.constraints.*;import kr.releasepilot.controlplane.identity.UserAccountPrincipal;import org.springframework.http.*;import org.springframework.security.access.prepost.PreAuthorize;import org.springframework.security.core.annotation.AuthenticationPrincipal;import org.springframework.web.bind.annotation.*;import java.util.UUID;
@RestController @RequestMapping("/api/v1/releases/{releaseId}") @PreAuthorize("hasRole('OPERATOR')") public class RolloutOperationController{
 private final RolloutOperationService service;public RolloutOperationController(RolloutOperationService service){this.service=service;}
 @PostMapping("/promote")@ResponseStatus(HttpStatus.ACCEPTED)RolloutOperationService.Receipt promote(@PathVariable UUID releaseId,@RequestHeader("Idempotency-Key")@Size(min=16,max=128)String key,@Valid@RequestBody Reason r,@AuthenticationPrincipal UserAccountPrincipal actor){return service.request(releaseId,ArgoRolloutsGateway.Action.PROMOTE,r.reason(),key,actor.id());}
 @PostMapping("/pause")@ResponseStatus(HttpStatus.ACCEPTED)RolloutOperationService.Receipt pause(@PathVariable UUID releaseId,@RequestHeader("Idempotency-Key")@Size(min=16,max=128)String key,@Valid@RequestBody Reason r,@AuthenticationPrincipal UserAccountPrincipal actor){return service.request(releaseId,ArgoRolloutsGateway.Action.PAUSE,r.reason(),key,actor.id());}
 @PostMapping("/resume")@ResponseStatus(HttpStatus.ACCEPTED)RolloutOperationService.Receipt resume(@PathVariable UUID releaseId,@RequestHeader("Idempotency-Key")@Size(min=16,max=128)String key,@Valid@RequestBody Reason r,@AuthenticationPrincipal UserAccountPrincipal actor){return service.request(releaseId,ArgoRolloutsGateway.Action.RESUME,r.reason(),key,actor.id());}
 @PostMapping("/abort")@ResponseStatus(HttpStatus.ACCEPTED)RolloutOperationService.Receipt abort(@PathVariable UUID releaseId,@RequestHeader("Idempotency-Key")@Size(min=16,max=128)String key,@Valid@RequestBody Reason r,@AuthenticationPrincipal UserAccountPrincipal actor){return service.request(releaseId,ArgoRolloutsGateway.Action.ABORT,r.reason(),key,actor.id());}
 public record Reason(@NotBlank@Size(max=1000)String reason){}
}